+++ packages/page-docs/src/scripts/layout/Landing.tsx
@@ -3,11 +3,13 @@ import { Footer } from './Footer';
 import { Header } from './Header';
 import { TagLine } from '../components';
 
+const version = process.env.BUILD_PCKG_VERSION;
+
 export const Landing: React.SFC = ({ children }) => (
   <div className="landing-page body-green">
     <div className="page-wrapper">
       <Header centered>
-        <TagLine />
+        <TagLine version={version} />
       </Header>
       <div className="doc-wrapper">{children}</div>
     </div>