+++ packages/piral-core/src/components/default/errorInfo.tsx
@@ -11,12 +11,14 @@ export const DefaultErrorInfo: React.SFC<ErrorInfoProps> = props => (
       switch (props.type) {
         case 'not_found':
           return <div key="default_error">Page {props.location.pathname} not found.</div>;
-        case 'feed':
-          return <div key="default_error">Feed error: {props.error}.</div>;
         case 'page':
           return <div key="default_error">Page {props.location.pathname} crashed.</div>;
         case 'loading':
           return <div key="default_error">Page could not be loaded: {props.error}.</div>;
+        case 'feed':
+          return <div key="default_error">Feed error: {props.error}.</div>;
+        case 'form':
+          return <div key="default_error">Form error: {props.error}.</div>;
       }
     }}
   />