+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Piral Changelog
 
+## 0.3.0 (to be determined)
+
+* Documentation enhancements
+* Added project landing page
+* Integration tests for Piral CLI
+* Updated homepage
+* Included fetch and GraphQL API extensions in `piral-ext`
+* Changed `piral` to a framework (`piral-core` remains a library)
+* Implemented support for pilet split (#1)
+* Added version to documentation page (#16)
+
 ## 0.2.0 (April 4, 2019)
 
 * Further tooling improvements
@@ -7,7 +18,7 @@
 * Renamed `PortalApi` to `PiralApi`
 * Renamed `PortalInstance` to `PiralInstance`
 * Added `registerSearchProvider` API
-* * Added `createForm` API
+* Added `createForm` API
 * Implemented `pilet-upgrade` command (#11)
 * Renamed the forwarded API to `piral`